# openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/admin/revisions/{revisionId}/imports": {
      "post": {
        "operationId": "createSpatialImport",
        "parameters": [
          {
            "name": "revisionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file",
                  "mode",
                  "clientRequestId"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  },
                  "format": {
                    "type": "string",
                    "enum": [
                      "csv",
                      "xlsx",
                      "geojson",
                      "kml"
                    ]
                  },
                  "mode": {
                    "type": "string",
                    "enum": [
                      "append",
                      "replace",
                      "upsert"
                    ]
                  },
                  "clientRequestId": {
                    "type": "string",
                    "format": "uuid"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "data": {
                      "type": "object",
                      "required": [
                        "id",
                        "revisionId",
                        "status",
                        "format",
                        "mode",
                        "file",
                        "progress",
                        "counts"
                      ],
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "revisionId": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "status": {
                          "type": "string"
                        },
                        "format": {
                          "type": "string",
                          "enum": [
                            "csv",
                            "xlsx",
                            "geojson",
                            "kml"
                          ]
                        },
                        "mode": {
                          "type": "string",
                          "enum": [
                            "append",
                            "replace",
                            "upsert"
                          ]
                        },
                        "file": {
                          "type": "object",
                          "required": [
                            "name",
                            "sizeBytes"
                          ],
                          "properties": {
                            "name": {
                              "type": "string"
                            },
                            "sizeBytes": {
                              "type": "integer"
                            }
                          }
                        },
                        "progress": {
                          "type": "integer",
                          "minimum": 0,
                          "maximum": 100
                        },
                        "counts": {
                          "type": "object",
                          "additionalProperties": {
                            "type": "number"
                          }
                        },
                        "failureCode": {
                          "type": "string",
                          "nullable": true
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    },
                    "meta": {
                      "type": "object",
                      "required": [
                        "requestId"
                      ],
                      "properties": {
                        "requestId": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "adminSession": []
          }
        ],
        "summary": "",
        "tags": [
          "imports"
        ]
      }
    },
    "/api/v1/admin/imports/{importId}": {
      "get": {
        "operationId": "getSpatialImport",
        "parameters": [
          {
            "name": "importId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "data": {
                      "type": "object",
                      "required": [
                        "id",
                        "revisionId",
                        "status",
                        "format",
                        "mode",
                        "file",
                        "progress",
                        "counts"
                      ],
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "revisionId": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "status": {
                          "type": "string"
                        },
                        "format": {
                          "type": "string",
                          "enum": [
                            "csv",
                            "xlsx",
                            "geojson",
                            "kml"
                          ]
                        },
                        "mode": {
                          "type": "string",
                          "enum": [
                            "append",
                            "replace",
                            "upsert"
                          ]
                        },
                        "file": {
                          "type": "object",
                          "required": [
                            "name",
                            "sizeBytes"
                          ],
                          "properties": {
                            "name": {
                              "type": "string"
                            },
                            "sizeBytes": {
                              "type": "integer"
                            }
                          }
                        },
                        "progress": {
                          "type": "integer",
                          "minimum": 0,
                          "maximum": 100
                        },
                        "counts": {
                          "type": "object",
                          "additionalProperties": {
                            "type": "number"
                          }
                        },
                        "failureCode": {
                          "type": "string",
                          "nullable": true
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    },
                    "meta": {
                      "type": "object",
                      "required": [
                        "requestId"
                      ],
                      "properties": {
                        "requestId": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "adminSession": []
          }
        ],
        "summary": "",
        "tags": [
          "imports"
        ]
      }
    },
...
  },
  "info": {
    "title": "DanangMap API",
    "description": "DanangMap v2 spatial CMS contract",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "adminSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-danangmap_session"
      },
      "csrf": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token"
      }
    },
...
  }
}